allow None responses in list replies
If you use e.g. mget, and some of the keys have no values, it's important to
preserve the None values in the response so that you know which keys the
values correspond to.

So now mget("a", "b", "c") => (1, None, 4)
rather than                => (1, 4)

sort() changes as a consequence of this.